xai_visual_framework_full_release/requirements.txt
torch
torchvision
captum
matplotlib
Pillow
typer
pandas
nbformat
nbconvert
jinja2
streamlit
plotly